core: Reorganize packages
