+++ modules/openai4s-core/shared/src/main/scala-2/openai4s/chat/Chat.scala
@@ -1,4 +1,4 @@
-package openai4s.request
+package openai4s.chat
 
 import cats.data.NonEmptyList
 import cats.{Eq, Show}
