+++ modules/openai4s-core/shared/src/main/scala-2/openai4s/chat/Response.scala
@@ -1,4 +1,4 @@
-package openai4s.response
+package openai4s.chat
 
 import cats.{Eq, Show}
 import eu.timepit.refined.cats.*
